feat: rename `utils` directory to `loaders`
